backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
ffmpeg-python==0.2.0
openai-whisper==20231117
pydub==0.25.1

# File handling
python-multipart==0.0.6
aiofiles==24.1.0

# Optional: For development
python-dotenv==1.0.0

# AI/ML dependencies
sentence-transformers==2.2.2
librosa>=0.10.0
spacy>=3.7.0
